--- src/components/atoms/RecipePreviewPicture/RecipePreviewPicture.stories.ts
@@ -1,19 +1,0 @@
-import { RecipePreviewPicture } from "./RecipePreviewPicture"
-
-import type { Meta, StoryObj } from "@storybook/react"
-
-const meta: Meta<typeof RecipePreviewPicture> = {
-  title: "Atoms/RecipePreviewPicture",
-  component: RecipePreviewPicture,
-  tags: ["autodocs"],
-}
-
-export default meta
-type Story = StoryObj<typeof RecipePreviewPicture>
-
-export const Normal: Story = {
-  args: {
-    index: 1,
-    imagePath: "sample-meal.png",
-  },
-}
